py/crawler/requirement.txt
Pillow
numpy
paddlehub
opencv-python